+++ data/resources/CHANGELOG.md
@@ -1,3 +1,8 @@
+December 7, 2021:
+- Happy Terror of Luclin Day!
+- Added full support for the merchant perk and bag slot 11 and 12.
+- Added ${Me.NumBagSlots} which returns the number of bag slots enabled in main inventory.
+
 November 30, 2021:
 - MacroQuest tray utility will now check Application Compatibility layers on startup.  To disable this check set
   DisableAppCompatCheck=true in the MacroQuest section of your MacroQuest.ini